examples/MINST_RAW_format/requirements.txt
tensorflow==2.2.0
kafka-python==2.0.1